requirements.txt
django==2.1
djangorestframework
django-cors-middleware
requests
